fix: make sure pg_tests run
pg_test feature gated tests did not run as part of the supabase-wrappers crate
because it's a cdylib and doesn't have Postgres running when the test run. Such
tests needed a real Postgres instance available to run them, which is run by
pgrx for an actual extension not a cdylib. Moved pg_test gated tests from
supabase-wrappers crate into the wrappers crate.
