+++ supabase-wrappers/Cargo.toml
@@ -13,13 +13,12 @@ rust-version = { workspace = true }
 
 [features]
 default = ["pg15"]
-pg13 = ["pgrx/pg13", "pgrx-tests/pg13"]
-pg14 = ["pgrx/pg14", "pgrx-tests/pg14"]
-pg15 = ["pgrx/pg15", "pgrx-tests/pg15"]
-pg16 = ["pgrx/pg16", "pgrx-tests/pg16"]
-pg17 = ["pgrx/pg17", "pgrx-tests/pg17"]
-pg18 = ["pgrx/pg18", "pgrx-tests/pg18"]
-pg_test = []
+pg13 = ["pgrx/pg13"]
+pg14 = ["pgrx/pg14"]
+pg15 = ["pgrx/pg15"]
+pg16 = ["pgrx/pg16"]
+pg17 = ["pgrx/pg17"]
+pg18 = ["pgrx/pg18"]
 
 [dependencies]
 pgrx = { version = "=0.19.2", default-features = false }
@@ -27,9 +26,6 @@ thiserror = "1.0.63"
 tokio = { version = "1.43", features = ["rt", "net"] }
 uuid = { version = "1.10.0" }
 supabase-wrappers-macros = { version = "0.1", path = "../supabase-wrappers-macros" }
-
-[dev-dependencies]
-pgrx-tests = "=0.19.2"
 
 [package.metadata.docs.rs]
 features = ["pg15"]
